refactor(config): expose agent preset resolution below agents backend
Move pure agent preset/config resolution functions from
agents/backends/codeagent_config.py to new config/agent_preset.py module.

This eliminates illegal upward dependencies:
- services/actor_support.py no longer imports vibe3.agents.*
- execution/execution_role_policy.py imports config resolution from config layer

Backend-specific sync_models_json() remains in agents/backends/codeagent_config.py.

Functions moved:
- resolve_effective_agent_options
- resolve_repo_agent_preset
- resolve_repo_agent_preset_name
- has_agent_env_override
- repo_models_json_path
- _read_models_json
- configured_backends
- find_missing_backend_commands
- Constants: REPO_MODELS_JSON_PATH, LEGACY_REPO_MODELS_JSON_PATH, BACKEND_COMMANDS

All import statements and test patch targets updated to use new location.
Path calculation adjusted (parents[3] instead of parents[4]) due to module depth change.

Tests: 72 targeted tests pass
Types: mypy passes on all affected files
