+++ src/vibe3/server/app.py
@@ -12,8 +12,8 @@ from typing import Annotated
 import typer
 import uvicorn
 
-from vibe3.agents.backends.codeagent_config import find_missing_backend_commands
 from vibe3.clients.git_client import GitClient
+from vibe3.config.agent_preset import find_missing_backend_commands
 from vibe3.config.orchestra_settings import load_orchestra_config
 from vibe3.models.orchestra_config import OrchestraConfig
 from vibe3.observability.logger import setup_logging
